refactor: split process spawn handler into focused modules
Move process:spawn orchestration out of process.ts into process/handle-spawn.ts
and phase helpers while keeping registerProcessHandlers as the IPC entry point.

Extract spawn-types, Claude transcript sanitization, Claude context resolution,
local interactive maestro-p realization, session mode persistence, and SSH
wrapping into dedicated modules. Route local maestro-p spawns through
applyClaudeSpawnDecision so desktop matches Cue, group chat, and tab naming.

Preserve all spawn-path comments and existing process.test.ts coverage (98 tests).
Other process IPC handlers (write, kill, runCommand, etc.) remain in process.ts.
